+++ src/ioabc.cpp
@@ -778,7 +778,7 @@ void ABCInput::parseTempo(const std::string &tempoString)
     Tempo *tempo = new Tempo();
     if (tempoString.find('=') != std::string::npos) {
         const int numStart = int(tempoString.find('=') + 1);
-        tempo->SetMm(std::atoi(tempoString.substr(numStart).c_str()));
+        tempo->SetMm(std::atof(tempoString.substr(numStart).c_str()));
     }
     if (tempoString.find('\"') != std::string::npos) {
         std::string tempoWord = tempoString.substr(tempoString.find('\"') + 1);
